AI-Grid-Infrastructure/requirements.txt
# AI Grid Orchestrator — Python Dependencies

# Core framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
streamlit>=1.33.0

# HTTP clients
httpx>=0.27.0
requests>=2.31.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Environment / config
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# Dev / testing
pytest>=8.1.0
pytest-asyncio>=0.23.0
